actualizar

+++ ejercicio_6.py
@@ -1,7 +1,7 @@
 print("Quiere saber cuanto cobra!!")
 horas=int(input("Ingrese la cantidad de horas trabajadas: "))
 
-costo=int(input("Ingrese el valor de la hora de trabajo: $"))
+costo=float(input("Ingrese el valor de la hora de trabajo: $"))
 pago=costo*horas
 
 print("el sueldo que debes cobrar es de: $",pago)
